backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
python-dotenv
razorpay
tensorflow
numpy
psycopg2-binary
pandas
scikit-learn
